+++ magma-func-supp/src/test/java/eu/lunisolar/magma/func/supp/delta/LFloatBinaryOpDeltaTest.java
@@ -36,7 +36,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 
 @SuppressWarnings("UnusedDeclaration")
-public class LFloatBinaryOpDeltaTest {
+public class LFloatBinaryOpDeltaTest  {
 
     private int functionCallCount = 0;
     private final float initialTestValue = 1f;
@@ -55,7 +55,7 @@ public class LFloatBinaryOpDeltaTest {
         assertThat(sut.lastValue())
             .isEqualTo(initialLastValue);
 
-        assertThat(sut.doApplyAsFloat((float)100,(float)100))
+        assertThat(sut.doApplyAsFloat(100f,100f))
             .isEqualTo(testValue);
         assertThat(functionCallCount).isEqualTo(1);
 
@@ -69,7 +69,7 @@ public class LFloatBinaryOpDeltaTest {
             .isEqualTo(initialTestValue);
         assertThat(functionCallCount).isEqualTo(1);
 
-        assertThat(sut.doApplyAsFloat((float)100,(float)100))
+        assertThat(sut.doApplyAsFloat(100f,100f))
             .isEqualTo(9f);
 
         assertThat(functionCallCount).isEqualTo(2);
