+++ magma-func-supp/src/test/java/eu/lunisolar/magma/func/supp/delta/LIntBinaryOpDeltaTest.java
@@ -36,7 +36,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 
 @SuppressWarnings("UnusedDeclaration")
-public class LIntBinaryOpDeltaTest {
+public class LIntBinaryOpDeltaTest  {
 
     private int functionCallCount = 0;
     private final int initialTestValue = 1;
@@ -55,7 +55,7 @@ public class LIntBinaryOpDeltaTest {
         assertThat(sut.lastValue())
             .isEqualTo(initialLastValue);
 
-        assertThat(sut.doApplyAsInt((int)100,(int)100))
+        assertThat(sut.doApplyAsInt(100,100))
             .isEqualTo(testValue);
         assertThat(functionCallCount).isEqualTo(1);
 
@@ -69,7 +69,7 @@ public class LIntBinaryOpDeltaTest {
             .isEqualTo(initialTestValue);
         assertThat(functionCallCount).isEqualTo(1);
 
-        assertThat(sut.doApplyAsInt((int)100,(int)100))
+        assertThat(sut.doApplyAsInt(100,100))
             .isEqualTo(9);
 
         assertThat(functionCallCount).isEqualTo(2);
