+++ magma-func-supp/src/test/java/eu/lunisolar/magma/func/supp/delta/LLongBinaryOpDeltaTest.java
@@ -36,7 +36,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 
 @SuppressWarnings("UnusedDeclaration")
-public class LLongBinaryOpDeltaTest {
+public class LLongBinaryOpDeltaTest  {
 
     private int functionCallCount = 0;
     private final long initialTestValue = 1L;
@@ -55,7 +55,7 @@ public class LLongBinaryOpDeltaTest {
         assertThat(sut.lastValue())
             .isEqualTo(initialLastValue);
 
-        assertThat(sut.doApplyAsLong((long)100,(long)100))
+        assertThat(sut.doApplyAsLong(100L,100L))
             .isEqualTo(testValue);
         assertThat(functionCallCount).isEqualTo(1);
 
@@ -69,7 +69,7 @@ public class LLongBinaryOpDeltaTest {
             .isEqualTo(initialTestValue);
         assertThat(functionCallCount).isEqualTo(1);
 
-        assertThat(sut.doApplyAsLong((long)100,(long)100))
+        assertThat(sut.doApplyAsLong(100L,100L))
             .isEqualTo(9L);
 
         assertThat(functionCallCount).isEqualTo(2);
